requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
datasets==3.2.0
dill==0.3.8
einops==0.8.0
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.3.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
keras==3.5.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
modelscope==1.22.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
networkx==3.3
numpy==1.26.4
open_clip_torch==2.30.0
opencv-python==4.11.0.86
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
propcache==0.2.1
protobuf==4.25.4
pyarrow==19.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-version==0.0.2
pytz==2024.2
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.8.0
safetensors==0.4.4
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.3.1
simplejson==3.19.3
six==1.16.0
sortedcontainers==2.4.0
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
timm==1.0.14
tokenizers==0.19.1
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchvision==0.19.0+cu118
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yarl==1.18.3
